+++ aws-java-sdk-accessanalyzer/src/main/java/com/amazonaws/services/accessanalyzer/model/transform/AnalyzerSummaryJsonUnmarshaller.java
@@ -68,6 +68,14 @@ public class AnalyzerSummaryJsonUnmarshaller implements Unmarshaller<AnalyzerSum
                     context.nextToken();
                     analyzerSummary.setName(context.getUnmarshaller(String.class).unmarshall(context));
                 }
+                if (context.testExpression("status", targetDepth)) {
+                    context.nextToken();
+                    analyzerSummary.setStatus(context.getUnmarshaller(String.class).unmarshall(context));
+                }
+                if (context.testExpression("statusReason", targetDepth)) {
+                    context.nextToken();
+                    analyzerSummary.setStatusReason(StatusReasonJsonUnmarshaller.getInstance().unmarshall(context));
+                }
                 if (context.testExpression("tags", targetDepth)) {
                     context.nextToken();
                     analyzerSummary.setTags(new MapUnmarshaller<String, String>(context.getUnmarshaller(String.class), context.getUnmarshaller(String.class))